+++ src/main/java/org/openrewrite/java/spring/security6/RemoveOauth2LoginConfig.java
@@ -15,32 +15,22 @@
  */
 package org.openrewrite.java.spring.security6;
 
-import org.openrewrite.Cursor;
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Recipe;
 import org.openrewrite.TreeVisitor;
 import org.openrewrite.internal.lang.Nullable;
-import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.MethodMatcher;
 import org.openrewrite.java.search.UsesMethod;
-import org.openrewrite.java.tree.J;
+import org.openrewrite.java.spring.RemoveMethodInvocationsVisitor;
 
 import java.time.Duration;
+import java.util.ArrayList;
+import java.util.List;
 
 public class RemoveOauth2LoginConfig extends Recipe {
-
     private static final MethodMatcher O_AUTH_2_LOGIN_MATCHER = new MethodMatcher(
         "org.springframework.security.config.annotation.web.builders.HttpSecurity oauth2Login()"
     );
-
-    private static final MethodMatcher USER_INFO_ENDPOINT_MATCHER = new MethodMatcher(
-        "org.springframework.security.config.annotation.web.configurers.oauth2.client.OAuth2LoginConfigurer userInfoEndpoint()"
-    );
-
-    private static final MethodMatcher USER_AUTHORITIES_MAPPER_MATCHER = new MethodMatcher(
-        "org.springframework.security.config.annotation.web.configurers.oauth2.client.OAuth2LoginConfigurer$UserInfoEndpointConfig userAuthoritiesMapper(org.springframework.security.core.authority.mapping.GrantedAuthoritiesMapper)"
-    );
-
 
     @Override
     public String getDisplayName() {
@@ -80,38 +70,10 @@ public class RemoveOauth2LoginConfig extends Recipe {
 
     @Override
     protected TreeVisitor<?, ExecutionContext> getVisitor() {
-        return new JavaIsoVisitor<ExecutionContext>() {
-            @Override
-            public J.MethodInvocation visitMethodInvocation(J.MethodInvocation method,
-                                                            ExecutionContext executionContext) {
-                Cursor parent = getCursor().dropParentUntil(p -> p instanceof J.Assignment ||
-                                                                 p instanceof J.VariableDeclarations.NamedVariable ||
-                                                                 p instanceof J.Return ||
-                                                                 p instanceof J.Block ||
-                                                                 p instanceof J.CompilationUnit);
-                boolean isAssignment = parent.getValue() instanceof J.Assignment ||
-                                       parent.getValue() instanceof J.VariableDeclarations.NamedVariable ||
-                                       parent.getValue() instanceof J.Return;
-                if (isAssignment) {
-                    // don't rewrite if this method invocation is assigned to a variable.
-                    return method;
-                }
-
-                if (USER_AUTHORITIES_MAPPER_MATCHER.matches(method) ||
-                    O_AUTH_2_LOGIN_MATCHER.matches(method) ||
-                    USER_INFO_ENDPOINT_MATCHER.matches(method)) {
-
-                    if (method.getSelect() instanceof J.Identifier) {
-                        // remove this statement
-                        return null;
-                    } else if (method.getSelect() instanceof J.MethodInvocation) {
-                        // update the statement
-                        doAfterVisit(this);
-                        return method.getSelect().withPrefix(method.getPrefix());
-                    }
-                }
-                return method;
-            }
-        };
+        List<String> methods = new ArrayList<>();
+        methods.add("org.springframework.security.config.annotation.web.builders.HttpSecurity oauth2Login()");
+        methods.add("org.springframework.security.config.annotation.web.configurers.oauth2.client.OAuth2LoginConfigurer userInfoEndpoint()");
+        methods.add("org.springframework.security.config.annotation.web.configurers.oauth2.client.OAuth2LoginConfigurer$UserInfoEndpointConfig userAuthoritiesMapper(org.springframework.security.core.authority.mapping.GrantedAuthoritiesMapper)");
+        return new RemoveMethodInvocationsVisitor(methods);
     }
 }
